+++ src/chains/Base.sol
@@ -24,9 +24,13 @@ import {
     IPancakeSwapV3Callback
 } from "../core/univ3forks/PancakeSwapV3.sol";
 import {sushiswapV3Factory, sushiswapV3ForkId} from "../core/univ3forks/SushiswapV3.sol";
+import {
+    solidlyV3Factory, solidlyV3InitHash, solidlyV3ForkId, ISolidlyV3Callback
+} from "../core/univ3forks/SolidlyV3.sol";
 import {aerodromeFactory, aerodromeInitHash, aerodromeForkId} from "../core/univ3forks/AerodromeSlipstream.sol";
 import {alienBaseV3Factory, alienBaseV3ForkId} from "../core/univ3forks/AlienBaseV3.sol";
 import {baseXFactory, baseXForkId} from "../core/univ3forks/BaseX.sol";
+import {swapBasedV3Factory, swapBasedV3ForkId} from "../core/univ3forks/SwapBasedV3.sol";
 
 // Solidity inheritance is stupid
 import {SettlerAbstract} from "../SettlerAbstract.sol";
@@ -76,6 +80,10 @@ abstract contract BaseMixin is FreeMemory, SettlerBase, Velodrome {
             factory = sushiswapV3Factory;
             initHash = uniswapV3InitHash;
             callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
+        } else if (forkId == solidlyV3ForkId) {
+            factory = solidlyV3Factory;
+            initHash = solidlyV3InitHash;
+            callbackSelector = uint32(ISolidlyV3Callback.solidlyV3SwapCallback.selector);
         } else if (forkId == aerodromeForkId) {
             factory = aerodromeFactory;
             initHash = aerodromeInitHash;
@@ -88,6 +96,10 @@ abstract contract BaseMixin is FreeMemory, SettlerBase, Velodrome {
             factory = baseXFactory;
             initHash = uniswapV3InitHash;
             callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
+        } else if (forkId == swapBasedV3ForkId) {
+            factory = swapBasedV3Factory;
+            initHash = pancakeSwapV3InitHash;
+            callbackSelector = uint32(IPancakeSwapV3Callback.pancakeV3SwapCallback.selector);
         } else {
             revert UnknownForkId(forkId);
         }